requirements.txt
torch>=1.9.0
torchvision>=0.10.0
flwr>=1.4.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
PyYAML>=6.0